+++ Assets/RevCore/Audio/Tests/AudioManagerTests.cs
@@ -1,0 +1,23 @@
+using System.Reflection;
+using NUnit.Framework;
+using UnityEngine;
+
+namespace RevCore.Tests
+{
+    public class AudioManagerTests
+    {
+        [Test]
+        public void OnDestroy_DestroyedInstance_ClearsSingleton()
+        {
+            var go = new GameObject("AudioManager");
+            var manager = go.AddComponent<AudioManager>();
+            typeof(AudioManager).GetMethod("Awake", BindingFlags.Instance | BindingFlags.NonPublic).Invoke(manager, null);
+
+            Assert.AreSame(manager, AudioManager.Instance);
+
+            Object.DestroyImmediate(go);
+
+            Assert.IsNull(AudioManager.Instance);
+        }
+    }
+}